requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Async tasks
celery==5.3.6
kombu==5.3.4

# AI / Vision
ultralytics==8.1.0
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.26.4

# Utils
python-decouple==3.8
drf-spectacular==0.27.1

# Dev/test
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
